#17077. Commit services changes.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@24089 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ services/src/main/java/ish/oncourse/webservices/replication/v4/builders/InvoiceStubBuilder.java
@@ -9,7 +9,7 @@ public class InvoiceStubBuilder extends AbstractWillowStubBuilder<Invoice, Invoi
 	protected InvoiceStub createFullStub(Invoice entity) {
 
 		InvoiceStub stub = new InvoiceStub();
-		stub.setAmountOwing(entity.getAmountOwing());
+		stub.setAmountOwing(entity.getAmountOwing().toBigDecimal());
 		stub.setBillToAddress(entity.getBillToAddress());
 		stub.setContactId(entity.getContact().getId());
 		stub.setCreated(entity.getCreated());
@@ -23,8 +23,8 @@ public class InvoiceStubBuilder extends AbstractWillowStubBuilder<Invoice, Invoi
 		stub.setPublicNotes(entity.getPublicNotes());
 		stub.setShippingAddress(entity.getShippingAddress());
 		stub.setSource(entity.getSource().getDatabaseValue());
-		stub.setTotalExGst(entity.getTotalExGst());
-		stub.setTotalGst(entity.getTotalGst());
+		stub.setTotalExGst(entity.getTotalExGst().toBigDecimal());
+		stub.setTotalGst(entity.getTotalGst().toBigDecimal());
 		
 		return stub;
 	}
